Implement basic TCP server functionality
This commit updates the server to accept connections, receive requests, and send responses. This is done by introducing the usage of Socket class for client isolation. It also makes use of a constant response string "+PONG\r\n" to be sent back to the clients. These changes set up the basic functionality of a TCP server.