Refactor RespResponse and enhance Set command handling
Refactored the `GetPrefix` in RespResponse.cs using switch expression instead of the statement. Enhanced the handling of Set command in RespCommandFactory.cs where the set expiration event is now thoroughly logged. The expiration handling is now done in milliseconds in ExpiredTasks.cs. Also, debugging was enabled in the codecrafters.yml.