Add support for Redis slave role configuration
The Redis config now accepts master host and port details. These are set in RedisServer based on RedisConfig. The OptionsParser checks the supplied arguments for a '--replicaof' flag and sets the config role to slave and parses the necessary details accordingly. In this slave role, an effort to ping the master on server start has also been implemented.