Add ReplConf command to Redis server
A new command, ReplConf, is added, which includes changes from the enumeration RespCommandType to the Redis server. A new class, ReplConfCommand, is also created within the Commands directory to define the behavior of the ReplConf command. The ReplConf command is executed whenever the role of Redis server is a slave.